+++ jkorm/src/main/kotlin/com/tang/jkorm/paginate/OrderItem.kt
@@ -1,10 +1,8 @@
 package com.tang.jkorm.paginate
 
 import com.tang.jkorm.function.SFunction
-import com.tang.jkorm.utils.Fields
-import com.tang.jkorm.utils.Reflects
+import com.tang.jkorm.utils.Reflects.getColumnName
 import kotlin.reflect.KMutableProperty1
-import kotlin.reflect.jvm.javaField
 
 /**
  * Order item for paginate order by
@@ -23,14 +21,12 @@ class OrderItem<T> {
     }
 
     constructor(property: KMutableProperty1<T, *>, asc: Boolean = true) {
-        val filed = property.javaField!!
-        this.column = Reflects.getColumnName(filed)
+        this.column = getColumnName(property)
         this.asc = asc
     }
 
     constructor(function: SFunction<T, *>, asc: Boolean = true) {
-        val filed = Fields.getField(function)
-        this.column = Reflects.getColumnName(filed)
+        this.column = getColumnName(function)
         this.asc = asc
     }