+++ jkorm/src/main/kotlin/com/tang/jkorm/wrapper/where/comparison/AbstractComparisonWrapper.kt
@@ -1,6 +1,7 @@
 package com.tang.jkorm.wrapper.where.comparison
 
 import com.tang.jkorm.function.SFunction
+import com.tang.jkorm.paginate.OrderItem
 import com.tang.jkorm.utils.Reflects.getColumnName
 import com.tang.jkorm.wrapper.Wrapper
 import com.tang.jkorm.wrapper.enumeration.ComparisonOperator
@@ -9,6 +10,7 @@ import com.tang.jkorm.wrapper.statement.ComparisonStatement
 import com.tang.jkorm.wrapper.statement.LogicalStatement
 import com.tang.jkorm.wrapper.where.WhereBuilder
 import com.tang.jkorm.wrapper.where.WhereGroupByWrapper
+import com.tang.jkorm.wrapper.where.WhereOrderByWrapper
 import java.util.function.Consumer
 import kotlin.reflect.KMutableProperty1
 
@@ -27,8 +29,14 @@ abstract class AbstractComparisonWrapper<T, R, W>(
 
     protected lateinit var whereGroupByWrapper: WhereGroupByWrapper<T, R, W>
 
-    fun isGroupByInitialized(): Boolean {
+    protected lateinit var whereOrderByWrapper: WhereOrderByWrapper<T, R, W>
+
+    protected fun isGroupByInitialized(): Boolean {
         return this::whereGroupByWrapper.isInitialized
+    }
+
+    protected fun isOrderByInitialized(): Boolean {
+        return this::whereOrderByWrapper.isInitialized
     }
 
     private fun setLastLogicalOperator(logicalOperator: LogicalOperator) {
@@ -1598,7 +1606,7 @@ abstract class AbstractComparisonWrapper<T, R, W>(
      * @return WhereGroupByWrapper<T, R, W>
      */
     fun groupBy(vararg columns: String): WhereGroupByWrapper<T, R, W> {
-        whereGroupByWrapper = WhereGroupByWrapper(wrapper, columns.toMutableList())
+        whereGroupByWrapper = WhereGroupByWrapper(wrapper, this, columns.toMutableList())
         return whereGroupByWrapper
     }
 
@@ -1608,6 +1616,7 @@ abstract class AbstractComparisonWrapper<T, R, W>(
      * @param columns columns
      * @return WhereGroupByWrapper<T, R, W>
      */
+    @SafeVarargs
     fun groupBy(vararg columns: KMutableProperty1<T, *>): WhereGroupByWrapper<T, R, W> {
         return groupBy(*columns.map { getColumnName(it) }.toTypedArray())
     }
@@ -1618,8 +1627,54 @@ abstract class AbstractComparisonWrapper<T, R, W>(
      * @param columns columns
      * @return WhereGroupByWrapper<T, R, W>
      */
+    @SafeVarargs
     fun groupBy(vararg columns: SFunction<T, *>): WhereGroupByWrapper<T, R, W> {
         return groupBy(*columns.map { getColumnName(it) }.toTypedArray())
+    }
+
+    /**
+     * Order by operation
+     *
+     * @param orderBys order by items
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    @SafeVarargs
+    fun orderBy(vararg orderBys: OrderItem<T>): WhereOrderByWrapper<T, R, W> {
+        whereOrderByWrapper = WhereOrderByWrapper(wrapper, this, orderBys.toMutableList())
+        return whereOrderByWrapper
+    }
+
+    /**
+     * Order by operation
+     *
+     * @param column column name
+     * @param asc asc or desc
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderBy(column: String, asc: Boolean = true): WhereOrderByWrapper<T, R, W> {
+        return orderBy(OrderItem<T>(column, asc))
+    }
+
+    /**
+     * Order by operation
+     *
+     * @param column column property
+     * @param asc asc or desc
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderBy(column: KMutableProperty1<T, *>, asc: Boolean = true): WhereOrderByWrapper<T, R, W> {
+        return orderBy(OrderItem<T>(column, asc))
+    }
+
+    /**
+     * Order by operation
+     *
+     * @param column column function
+     * @param asc asc or desc
+     * @return WhereOrderByWrapper<T, R, W>
+     */
+    fun orderBy(column: SFunction<T, *>, asc: Boolean = true): WhereOrderByWrapper<T, R, W> {
+        return orderBy(OrderItem<T>(column, asc))
     }
 
 }