+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Support more comparison operators
  * Support return generated key for insert statement
  * Support group by for where condition
+ * Support order by for where condition
 
 ## 1.0.2